auth_service/auth_app/urls.py
from django.urls import path
from .views import register_user
from . import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from .views import get_nearby_landlords
urlpatterns = [
    path('register/', register_user, name='register'),
    path('login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('tenant/find_landlords/', get_nearby_landlords, name='find_landlords'),
]